backend/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
pydantic-settings==2.11.0
